Use BigInteger for user ids and foreign keys

+++ Frendly_Map/bot/models/faq_entry.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Text, DateTime, ForeignKey
+from sqlalchemy import BigInteger, Column, Integer, String, Text, DateTime, ForeignKey
 from sqlalchemy.sql import func
 
 from bot.models.base import Base
@@ -10,6 +10,6 @@ class FaqEntry(Base):
     id = Column(Integer, primary_key=True, index=True)
     question = Column(String, nullable=False)
     answer = Column(Text, nullable=False)
-    created_by = Column(Integer, ForeignKey("users.id"), nullable=True)
+    created_by = Column(BigInteger, ForeignKey("users.id"), nullable=True)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     updated_at = Column(DateTime(timezone=True), onupdate=func.now())
