+++ Frendly_Map/bot/models/support_ticket.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, ForeignKey, Text
+from sqlalchemy import BigInteger, Column, Integer, String, DateTime, ForeignKey, Text
 from sqlalchemy.sql import func
 from sqlalchemy.orm import relationship
 
@@ -9,12 +9,12 @@ class SupportTicket(Base):
     __tablename__ = "support_tickets"
 
     id = Column(Integer, primary_key=True, index=True)
-    user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
+    user_id = Column(BigInteger, ForeignKey("users.id"), nullable=False)
     status = Column(String, default="open", nullable=False)
     subject = Column(String, nullable=True)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     closed_at = Column(DateTime(timezone=True), nullable=True)
-    closed_by = Column(Integer, ForeignKey("users.id"), nullable=True)
+    closed_by = Column(BigInteger, ForeignKey("users.id"), nullable=True)
     closed_reason = Column(Text, nullable=True)
 
     messages = relationship(
@@ -27,7 +27,7 @@ class SupportMessage(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     ticket_id = Column(Integer, ForeignKey("support_tickets.id"), nullable=False)
-    sender_id = Column(Integer, ForeignKey("users.id"), nullable=True)
+    sender_id = Column(BigInteger, ForeignKey("users.id"), nullable=True)
     sender_role = Column(String, nullable=False)
     message = Column(Text, nullable=False)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
